requirements.airflow.txt
# ============================================================================
# Adaptive Data Governance Framework - Airflow Container Dependencies
# ============================================================================
# Dependencies installed into the Airflow image so DAG task callables can
# import the framework's src modules.  The base apache/airflow image already
# provides airflow, psycopg2, etc.
# ============================================================================

# -- Spark & Delta (needed by every task that touches data) ------------------
pyspark==3.5.0
delta-spark==3.0.0

# -- Data Quality -----------------------------------------------------------
great-expectations==0.18.8
pydantic==2.5.2
pandera==0.17.2

# -- PII Detection ----------------------------------------------------------
presidio-analyzer==2.2.354
presidio-anonymizer==2.2.354

# -- NLP (used by PII detector) ---------------------------------------------
transformers==4.36.2
torch==2.1.2
sentencepiece==0.1.99

# -- Core utilities used across all src modules ------------------------------
pandas==2.1.4
numpy==1.26.2
pyarrow==14.0.1
faker==21.0.0
loguru==0.7.2
pyyaml==6.0.1
python-dotenv==1.0.0
tqdm==4.66.1
scikit-learn==1.3.2

# -- Visualization (used by data_profiler & dq_framework) -------------------
matplotlib==3.8.2
seaborn==0.13.1
plotly==5.18.0
